Properly define Notification.get_or_create

+++ app/slogan/signals.py
@@ -1,4 +1,4 @@
-from django.db.models.signals import post_save, post_delete, pre_save, pre_delete
+from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.contrib.auth.models import User
 from django.core.urlresolvers import reverse
@@ -12,12 +12,12 @@ from .models import Slogan
 @receiver(post_save, sender=Slogan)
 def new_slogan_notification(sender, created, **kwargs):
     if not kwargs.get('raw', False):  # Fixture?
-        if created:
-            slogan = kwargs['instance']
-            if not slogan.is_visible:
-                recipients = User.objects.filter(groups__name='SloganTeam')
+        slogan = kwargs['instance']
+        if created and not slogan.is_visible:
+            recipients = User.objects.filter(groups__name='SloganTeam')
+            for recipient in recipients:
                 Notification.objects.get_or_create(
-                        recipients=recipients,
+                        recipient=recipient,
                         title='Nouveau slogan',
                         description='Le slogan <em>%s</em> a été proposé par %s et doit être validé pour être visible.'
                                     % (force_escape(slogan.slogan), slogan.author),
@@ -29,7 +29,7 @@ def new_slogan_notification(sender, created, **kwargs):
 @receiver(post_save, sender=Slogan)
 def remove_notifications_when_approved(sender, created, **kwargs):
     if not kwargs.get('raw', False):  # Fixture?
-        if not created:
-            slogan = kwargs['instance']
+        slogan = kwargs['instance']
+        if not created and slogan.is_visible:
             if slogan.is_visible:
                 Notification.objects.filter(app='slogan', key=slogan.pk).delete()
